pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.11.0"]
build-backend = "uv_build"

[project]
name = "skin_type_classifier"
version = "0.0.1"
requires-python = ">=3.12"
description = "From Public Datasets to Real-World Deployment:  Robust Skin-Type Classification in Teledermatology"
authors = [
  { name = "Alex Serrano", email = "alem5005@gmail.com" },
]
readme = "README.md"
dependencies = [
    "fastapi==0.135.1",
    "hydra-core>=1.3",
    "matplotlib>=3.10.8",
    "omegaconf>=2.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "pytorch-triton-xpu>=3.5.0 ; sys_platform == 'win32'",
    "typer==0.24.1",
    "uvicorn==0.42.0",
]

[dependency-groups]
dev = [
    "coverage==7.13.5",
    "dvc>=3.66.1",
    "dvc-gs>=3.0.2",
    "ipykernel>=7.2.0",
    "mkdocs-material==9.7.5",
    "mkdocstrings-python==2.0.3",
    "mypy==1.19.1",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "ruff==0.15.6",
    "seaborn>=0.13.2",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-xpu", marker = "sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-xpu", marker = "sys_platform == 'win32'" },
]
pytorch-triton-xpu = [
  { index = "pytorch-xpu", marker = "sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-xpu"
url = "https://download.pytorch.org/whl/xpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Pure unit tests with no external data dependencies",
    "data: Tests that require data/processed/full_dataset.csv (DVC-tracked)",
]

[tool.coverage.run]
omit = ["tests/*"]
